Bunch of changes for v2 Commit some unfinished stuff because my mac is acting weird and i don't want to lose this.

+++ RyteBytes/StripeCustomer.m
@@ -20,8 +20,8 @@
     
     NSMutableDictionary *card = [[NSMutableDictionary alloc] init];
     [card setValue:creditCard.number forKey:@"number"];
-    [card setValue:[NSString stringWithFormat:@"%ld", creditCard.expMonth] forKey:@"exp_month"];
-    [card setValue:[NSString stringWithFormat:@"%ld", creditCard.expYear] forKey:@"exp_year"];
+    [card setValue:[NSString stringWithFormat:@"%ld", (unsigned long)creditCard.expMonth] forKey:@"exp_month"];
+    [card setValue:[NSString stringWithFormat:@"%ld", (unsigned long)creditCard.expYear] forKey:@"exp_year"];
     [card setValue:creditCard.cvc forKey:@"cvc"];
     
     return self;
